requirements.txt
streamlit
python-dotenv
langchain>=0.2.0
langchain-groq>=0.1.4
tiktoken
